requirements.txt
streamlit
youtube-transcript-api
faiss-cpu
langchain
langchain-community
huggingface-hub
python-dotenv
langchain_huggingface